Tests/LibWeb/Text/expected/HTML/align-presentational-hint.txt

object align=center: middle
embed align=center: middle
input image align=center: middle
object align=middle: middle
embed align=middle: middle
input image align=middle: middle
